refactor(efd-server): ♻️ loot quality in home realm has been reduced
